New features and updates.
1. Now you can select any audio track of your wish rather than the previous, only the most popular one.
2. Now there are two available dimensions for the artwork | Tall (2048 x 2732) and Square (3840 x 3840).

Previously I was using web scraping methodology and was also working fine but recently due to some HTML changes in the apple music web, the program was throwing an error. So now it is using AppleMusic web api.

You might face auth bearer token error in case if the token given in the script expires, I am trying to find a way to fetch the token automatically.